requirements.txt
Package                      Version
---------------------------- -----------
annotated-types              0.7.0
attrs                        24.2.0
cachetools                   5.5.0
certifi                      2024.8.30
cffi                         1.17.1
charset-normalizer           3.3.2
colorama                     0.4.6
google-ai-generativelanguage 0.6.9
google-api-core              2.19.2
google-api-python-client     2.145.0
google-auth                  2.34.0
google-auth-httplib2         0.2.0
google-generativeai          0.8.0
googleapis-common-protos     1.65.0
grpcio                       1.66.1
grpcio-status                1.66.1
h11                          0.14.0
httplib2                     0.22.0
idna                         3.8
outcome                      1.3.0.post0
pip                          22.3.1
proto-plus                   1.24.0
protobuf                     5.28.0
pyasn1                       0.6.1
pyasn1_modules               0.4.1
pycparser                    2.22
pydantic                     2.9.1
pydantic_core                2.23.3
pyparsing                    3.1.4
PySocks                      1.7.1
python-decouple              3.8
requests                     2.32.3
rsa                          4.9
selenium                     4.24.0
setuptools                   65.5.0
sniffio                      1.3.1
sortedcontainers             2.4.0
tqdm                         4.66.5
trio                         0.26.2
trio-websocket               0.11.1
typing_extensions            4.12.2
uritemplate                  4.1.1
urllib3                      2.2.2
websocket-client             1.8.0
wsproto                      1.2.0
